feat : LineTestFixture 생성
